+++ lib/market_hub/models/base.rb
@@ -11,9 +11,10 @@ module MarketHub
       attr_reader :client
       attr_reader :body
 
-      def initialize(client, body = {})
+      def initialize(client, body = {}, options = {})
         @client = client
         @body = body
+        @options = options
       end
 
       def save
@@ -21,6 +22,14 @@ module MarketHub
       end
 
       def destroy
+        raise MarketHub::NotImplemented, 'Sobreescreva este método na classe referente ao marketplace que você esta criando'
+      end
+
+      def active!
+        raise MarketHub::NotImplemented, 'Sobreescreva este método na classe referente ao marketplace que você esta criando'
+      end
+
+      def paused!
         raise MarketHub::NotImplemented, 'Sobreescreva este método na classe referente ao marketplace que você esta criando'
       end
 
@@ -34,13 +43,15 @@ module MarketHub
 
       def self.find(client, id)
         if client.connected?
-          self.new(client, client.find(id))
+          self.new(client, client.find(id), { action: 'find' })
         end
       end
 
       def self.create(client, body)
         if client.connected?
-          self.new(client, body).save
+          object = self.new(client, body, { action: 'create' })
+          object.save
+          object
         end
       end